final changes

+++ bamazonDB.sql
@@ -3,11 +3,11 @@ CREATE DATABASE bamazon;
 USE bamazon;
 
 CREATE TABLE products(
-  item_id INT NOT NULL AUTO_INCREMENT,
+  item_id INT(10) NOT NULL AUTO_INCREMENT,
   product_name VARCHAR(45) NOT NULL,
   department_name VARCHAR(45) NOT NULL,
   price INTEGER(10) NOT NULL,
-  stock_quantity INTEGER(10) NOT NULL,
+  stock_quantity INT(10) NOT NULL,
   PRIMARY KEY (item_id)
 );
